revised some files related to squeeze
