--- src/squeeze.js
@@ -1,20 +1,0 @@
-'use strict';
-
-import {reshape} from './reshape.js';
-import {validateSqueezeParams} from './lib/validate-input.js';
-
-/**
- * Reduce the rank of a tensor by eliminating dimensions with size 1 of the tensor shape.
- * @param {Tensor} input
- * @param {MLSqueezeOptions} options
- * @return {Tensor}
- */
-export function squeeze(input, {axes} = {}) {
-  validateSqueezeParams(...arguments);
-  const inpAxes = axes ?? new Array(input.rank).fill(0).map((_, i) => i);
-
-  const outputShape = input.shape.filter((dim, axis) =>
-    !(dim === 1 && inpAxes.indexOf(axis) !== -1));
-  const output = reshape(input, outputShape);
-  return output;
-}
